Fix numerical casting warnings
Refs #19094

+++ MantidQt/CustomInterfaces/src/Indirect/ApplyPaalmanPings.cpp
@@ -622,11 +622,11 @@ void ApplyPaalmanPings::plotInPreview(const QString &curveName,
   if (ws->getNumberHistograms() > m_spectra) {
     m_uiForm.ppPreview->addSpectrum(curveName, ws, m_spectra, curveColor);
   } else if (m_ppSampleWS || m_ppContainerWS) {
-    int specNo = std::min(m_ppContainerWS->getNumberHistograms(),
-                          m_ppSampleWS->getNumberHistograms()) -
-                 boost::numeric_cast<size_t>(1);
+    size_t specNo = std::min(m_ppContainerWS->getNumberHistograms(),
+                             m_ppSampleWS->getNumberHistograms()) -
+                             1;
     m_uiForm.ppPreview->addSpectrum(curveName, ws, specNo, curveColor);
-    m_uiForm.spPreviewSpec->setValue(specNo);
+    m_uiForm.spPreviewSpec->setValue(boost::numeric_cast<int>(specNo));
   }
 }
 } // namespace CustomInterfaces
